+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Fixed migration bug where Save Data setting was reset in migration
 - Fixed a bug where saving the Form settings redirected to an incorrect edit URL
 
+### Removed
+- Removed the fakeIt and isValid properties from the `OnBeforeValidateEntryEvent`
+
 ## 3.5.0 - 2019-10-10
 
 > {tip} IF: you upgrade, THEN: Conditional Fields! Projects using Custom Form Templates or extending Sprout Forms in other custom ways should read the upgrade notes before upgrading: [Changes in Sprout Forms v3.5.0](https://sprout.barrelstrengthdesign.com/docs/forms/installing-and-updating-craft-3.html#upgrading-to-forms-v3-5-0)